FEAT: added authorisation and documentation to endpoints

+++ MeFit.API/Controllers/ProfilesController.cs
@@ -10,10 +10,12 @@ using MeFit.DAL.Models.Domain;
 using AutoMapper;
 using MeFit.DAL.Models.DTOs.Profile;
 using System.Net.Mime;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.JsonPatch;
 
 namespace MeFit.API.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("api")]
     [ApiController]
     public class ProfilesController : ControllerBase
     {
@@ -25,22 +27,23 @@ namespace MeFit.API.Controllers
             _mapper = mapper;
         }
 
+        // GET: api/profile/profile_id
         /// <summary>
         /// Returns detail about current state of the users profile with their goals
         /// </summary>
         /// <param name="id">ID of a profile</param>
         /// <returns>Profile</returns>
         /// <response code="200">Returns a profile</response>
+        /// <response code="401">Not authorized</response>
         /// <response code="404">No profile found</response>
-        // GET: api/Profiles/profile_id
-        [HttpGet("{id}")]
+        [HttpGet("profile/{id}")]
+        //[Authorize]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<ProfileReadDTO>> GetProfileById([FromRoute] int id)
         {
-            //var profile = await _context.Profiles.FindAsync(id);
             var profile = await _context.Profiles.Include(p => p.Goals).FirstOrDefaultAsync(p => p.Id == id);
-            
+
             if (profile == null)
             {
                 return NoContent();
@@ -49,73 +52,83 @@ namespace MeFit.API.Controllers
             return Ok(profileReadDTO);
         }
 
+        // PATCH: api/profile/profile_id
         /// <summary>
         /// Executes partial update of the corresponding profile_id.
         /// </summary>
         /// <param name="id">ID of a profile</param>
-        /// <param name="profile"></param>
+        /// <param name="newProfile">Profiles new info</param>
         /// <response code="204">Successfully changed profile</response>
-        /// <response code="400">Bad request</response>
+        /// <response code="401">Not authorized</response>
         /// <response code="404">No profile found</response>
-        // PATCH: api/Profiles/profile_id
-        [HttpPatch("{id}")]
+        /// <response code="500">Internal Server Error</response>
+        [HttpPatch("profile/{id}")]
+        //[Authorize]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
-        public async Task<IActionResult> PutProfile([FromRoute] int id, [FromBody] ProfileUpdateDTO profile)
+        public async Task<IActionResult> UpdateProfile([FromRoute] int id, [FromBody] JsonPatchDocument<DAL.Models.Domain.Profile> newProfile)
         {
-            if (id != profile.Id)
+            var profile = await _context.Profiles.FindAsync(id);
+            if (profile == null)
             {
-                return BadRequest();
+                return NotFound();
             }
 
-            _context.Entry(profile).State = EntityState.Modified;
+            newProfile.ApplyTo(profile, ModelState);
 
             try
             {
                 await _context.SaveChangesAsync();
             }
-            catch (DbUpdateConcurrencyException)
+            catch
             {
-                if (!ProfileExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
+                return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
             return NoContent();
         }
 
+        // POST: api/profile
         /// <summary>
         /// Creates a new profile. Accepts appropriate parameters in the profile body as application/json
         /// </summary>
         /// <param name="newProfile">Profile info</param>
         /// <returns>A newly created profile</returns>
         /// <response code="201">Successfully created profile</response>
-        // POST: api/Profiles
-        [HttpPost]
-        [Consumes("application/json")]
+        /// <response code="401">Not authorized</response>
+        /// <response code="500">Internal Server Error</response>
+        [HttpPost("profile")]
+        //[Authorize]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status201Created)]
         public async Task<ActionResult<ProfileReadDTO>> PostProfile([FromBody] ProfileCreateDTO newProfile)
         {
             var domainProfile = _mapper.Map<DAL.Models.Domain.Profile>(newProfile);
             _context.Profiles.Add(domainProfile);
-            await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetProfile", new { id = domainProfile.Id }, newProfile);
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
+
+            return CreatedAtAction("GetProfileById", new { id = domainProfile.Id }, _mapper.Map<ProfileReadDTO>(domainProfile));
         }
 
+        // DELETE: api/profile/5 ------------- USER ONLY!!!!! -------------
         /// <summary>
-        /// Deletes a profile
+        /// Deletes a profile 
         /// </summary>
         /// <param name="id">ID of a profile</param>
         /// <response code="204">Successfully deleted profile</response>
+        /// <response code="401">Not authorized</response>
+        /// <response code="403">Not allowed(not having the necessary permissions)</response>
         /// <response code="404">No profile found</response>
-        // DELETE: api/Profiles/5 ------------- USER ONLY!!!!! -------------
-        [HttpDelete("{id}")]
+        /// <response code="500">Internal Server Error</response>
+        [HttpDelete("profile/{id}")]
+        //[Authorize]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<IActionResult> DeleteProfile(int id)
         {
@@ -125,15 +138,33 @@ namespace MeFit.API.Controllers
                 return NotFound();
             }
 
-            _context.Profiles.Remove(profile);
-            await _context.SaveChangesAsync();
+            var usernameFromToken = TakeUserNameFromToken();
+            if (profile.User.Id != TakeIdFromUser(usernameFromToken).Result)
+            {
+                return Forbid();
+            }
+
+            try
+            {
+                _context.Profiles.Remove(profile);
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return NoContent();
         }
-
-        private bool ProfileExists(int id)
+        private string TakeUserNameFromToken()
         {
-            return _context.Profiles.Any(e => e.Id == id);
+            var username = User.Claims.FirstOrDefault(c => c.Type == "preferred_username");
+            return username.Value;
+        }
+        private async Task<int> TakeIdFromUser(string usernameFromToken)
+        {
+            var id = (await _context.Users.FirstOrDefaultAsync(x => x.Username == usernameFromToken)).Id;
+            return id;
         }
     }
 }
